+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Allied reports now inform you when your team is the sole presence on a planet
+- Situation reports won't be issued by defeated allies
 
 ## v0.4 - 2023-04-27